// src/models/tmdb/person/PersonCrew.ts
export interface PersonCrew {
  adult: boolean;
  backdrop_path: string;
  genre_ids: Array<number>;
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
  department: string;
  job: string;
  credit_id: string;
  media_type: string;
}